♻️ Refactor: Break down App component from 627 to 231 lines
Addresses issue #38 - App component was too large and handled too many responsibilities.

## Changes

### Custom Hooks Created
- useFullscreen: Fullscreen management with vendor prefixes
- useTheme: Dark mode theme management
- useGameSettings: Game settings persistence to localStorage
- useGameState: Game state and navigation logic

### Components Created
- Header: App header with theme toggle, fullscreen, and auth controls
- Navigation: Navigation tabs for game screens
- GameRouter: Routes to appropriate game screen based on state
- AuthModal: Authentication modal
- ProfileModal: User profile modal

### Refactored App.tsx
- Reduced from 627 lines to 231 lines (63% reduction)
- Now acts as orchestrator using hooks and components
- Improved separation of concerns
- Better testability and maintainability

## Testing
- ✅ All existing tests pass (5/5)
- ✅ Production build successful
- ✅ No functionality changes - pure refactoring
