+++ src/hooks/useFullscreen.ts
@@ -1,0 +1,100 @@
+import { useState, useEffect, useCallback } from 'react';
+import { useError } from '../context/ErrorContext';
+
+/**
+ * Custom hook for managing fullscreen state and functionality
+ * Handles vendor prefixes for cross-browser compatibility
+ */
+export const useFullscreen = () => {
+  const { addError } = useError();
+  const [isFullScreen, setIsFullScreen] = useState(false);
+
+  // Listen for fullscreen changes
+  useEffect(() => {
+    const handleFullscreenChange = () => {
+      const isFullscreen = !!(
+        document.fullscreenElement ||
+        (document as any).webkitFullscreenElement ||
+        (document as any).mozFullScreenElement ||
+        (document as any).msFullscreenElement
+      );
+      setIsFullScreen(isFullscreen);
+    };
+
+    // Add event listeners for all vendor prefixes
+    document.addEventListener('fullscreenchange', handleFullscreenChange);
+    document.addEventListener('webkitfullscreenchange', handleFullscreenChange);
+    document.addEventListener('mozfullscreenchange', handleFullscreenChange);
+    document.addEventListener('MSFullscreenChange', handleFullscreenChange);
+
+    return () => {
+      document.removeEventListener('fullscreenchange', handleFullscreenChange);
+      document.removeEventListener(
+        'webkitfullscreenchange',
+        handleFullscreenChange
+      );
+      document.removeEventListener(
+        'mozfullscreenchange',
+        handleFullscreenChange
+      );
+      document.removeEventListener(
+        'MSFullscreenChange',
+        handleFullscreenChange
+      );
+    };
+  }, []);
+
+  const toggleFullscreen = useCallback(async () => {
+    try {
+      const doc = document as any;
+      const docEl = document.documentElement as any;
+
+      // Check if already in fullscreen (with vendor prefixes)
+      const isCurrentlyFullscreen = !!(
+        document.fullscreenElement ||
+        doc.webkitFullscreenElement ||
+        doc.mozFullScreenElement ||
+        doc.msFullscreenElement
+      );
+
+      if (!isCurrentlyFullscreen) {
+        // Request fullscreen with vendor prefix support
+        if (docEl.requestFullscreen) {
+          await docEl.requestFullscreen();
+        } else if (docEl.webkitRequestFullscreen) {
+          await docEl.webkitRequestFullscreen();
+        } else if (docEl.mozRequestFullScreen) {
+          await docEl.mozRequestFullScreen();
+        } else if (docEl.msRequestFullscreen) {
+          await docEl.msRequestFullscreen();
+        } else {
+          throw new Error('Fullscreen not supported');
+        }
+      } else {
+        // Exit fullscreen with vendor prefix support
+        if (document.exitFullscreen) {
+          await document.exitFullscreen();
+        } else if (doc.webkitExitFullscreen) {
+          await doc.webkitExitFullscreen();
+        } else if (doc.mozCancelFullScreen) {
+          await doc.mozCancelFullScreen();
+        } else if (doc.msExitFullscreen) {
+          await doc.msExitFullscreen();
+        }
+      }
+    } catch (error) {
+      console.warn('Fullscreen toggle failed:', error);
+      addError(
+        'Unable to toggle fullscreen. Your browser may not support this feature.'
+      );
+      // For iOS Safari, provide user feedback since fullscreen is very limited
+      if (/iPad|iPhone|iPod/.test(navigator.userAgent)) {
+        alert(
+          'Fullscreen not supported on this device. Try using "Add to Home Screen" for a fullscreen-like experience.'
+        );
+      }
+    }
+  }, [addError]);
+
+  return { isFullScreen, toggleFullscreen };
+};
